refactor(mcp): modularize Radar catalog tool
