+++ open-sse/mcp-server/radarCatalog.ts
@@ -1,4 +1,18 @@
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+
+import { logToolCall } from "./audit.ts";
+import { radarCatalogInput, radarCatalogOutput } from "./schemas/radarCatalog.ts";
+import type { McpToolExtraLike } from "./scopeEnforcement.ts";
+import type { TextToolResult } from "./toolResult.ts";
+import { sanitizeErrorMessage } from "../utils/error.ts";
+
 type JsonRecord = Record<string, unknown>;
+
+type ScopeEnforcer = (
+  toolName: string,
+  handler: (args: unknown, extra?: McpToolExtraLike) => Promise<TextToolResult>,
+  toolScopes?: readonly string[]
+) => (args: unknown, extra?: McpToolExtraLike) => Promise<TextToolResult>;
 
 export interface McpRadarCatalogArgs {
   provider?: string;
@@ -114,4 +128,43 @@ export async function getMcpRadarCatalog(
     .sort(compareEntries);
 
   return { meta: normalizeMeta(raw.meta), models };
+}
+
+async function handleRadarCatalog(args: {
+  provider?: string;
+  familyId?: string;
+  enabledOnly: boolean;
+}): Promise<TextToolResult> {
+  const start = Date.now();
+  try {
+    const result = radarCatalogOutput.parse(await getMcpRadarCatalog(args));
+    await logToolCall(
+      "omniroute_radar_catalog",
+      args,
+      { modelCount: result.models.length },
+      Date.now() - start,
+      true
+    );
+    return { content: [{ type: "text", text: JSON.stringify(result, null, 2) }] };
+  } catch (error) {
+    const message = sanitizeErrorMessage(error) || "Failed to read Radar catalog";
+    await logToolCall("omniroute_radar_catalog", args, null, Date.now() - start, false, message);
+    return { content: [{ type: "text", text: `Error: ${message}` }], isError: true };
+  }
+}
+
+export function registerRadarCatalogTool(
+  server: McpServer,
+  withScopeEnforcement: ScopeEnforcer
+): void {
+  server.registerTool(
+    "omniroute_radar_catalog",
+    {
+      description: "Reads the local signed Radar catalog with optional provider and family filters",
+      inputSchema: radarCatalogInput,
+    },
+    withScopeEnforcement("omniroute_radar_catalog", (args) =>
+      handleRadarCatalog(radarCatalogInput.parse(args))
+    )
+  );
 }
