+++ open-sse/mcp-server/schemas/tools.ts
@@ -13,11 +13,11 @@ import { z } from "zod";
 import { toolSearchTool } from "./toolSearch.ts";
 import { pickFastestModelTool } from "./pickFastestModel.ts";
 import { CCR_MCP_TOOLS } from "./ccrTools.ts";
+import { radarCatalogTool } from "./radarCatalog.ts";
 import {
   AUTO_ROUTING_STRATEGY_VALUES,
   ROUTING_STRATEGY_VALUES,
 } from "../../../src/shared/constants/routingStrategies.ts";
-
 // ============ Shared Types ============
 // AuditLevel + McpToolDefinition live in the leaf ./toolDefinition.ts so that
 // toolSearch.ts can import the type without forming a tools.ts ↔ toolSearch.ts cycle.
@@ -26,7 +26,6 @@ export type { AuditLevel, McpToolDefinition } from "./toolDefinition.ts";
 import type { McpToolDefinition } from "./toolDefinition.ts";
 export { pickFastestModelInput, pickFastestModelOutput } from "./pickFastestModel.ts";
 export * from "./ccrTools.ts";
-
 // ============ Phase 1: Essential Tools ============
 
 // --- Tool 1: omniroute_get_health ---
@@ -430,69 +429,6 @@ export const listModelsCatalogTool: McpToolDefinition<
   auditLevel: "none",
   phase: 1,
   sourceEndpoints: ["/api/models/catalog", "/v1/models"],
-};
-
-// --- Tool 9: omniroute_radar_catalog ---
-export const radarCatalogInput = z.object({
-  provider: z.string().trim().min(1).max(100).optional().describe("Filter by provider id"),
-  familyId: z.string().trim().min(1).max(120).optional().describe("Filter by curated family id"),
-  enabledOnly: z.boolean().default(true).describe("Exclude models disabled by the Radar feed"),
-});
-
-const radarLimitOutput = z.object({
-  rpm: z.number().nullable(),
-  rpd: z.number().nullable(),
-  tpm: z.number().nullable(),
-  tpd: z.number().nullable(),
-});
-
-export const radarCatalogOutput = z.object({
-  meta: z
-    .object({
-      version: z.string(),
-      tier: z.string(),
-      fetchedAt: z.string(),
-    })
-    .nullable(),
-  models: z.array(
-    z.object({
-      provider: z.string(),
-      modelId: z.string(),
-      displayName: z.string(),
-      familyId: z.string().nullable(),
-      quota: z.object({
-        monthlyTokens: z.number(),
-        creditTokens: z.number(),
-        freeType: z.string(),
-        limits: radarLimitOutput.nullable(),
-      }),
-      capabilities: z
-        .object({
-          tools: z.boolean(),
-          vision: z.boolean(),
-          thinking: z.boolean(),
-        })
-        .nullable(),
-      enabled: z.boolean(),
-      origin: z.enum(["baseline", "radar", "local"]),
-      disabledBy: z.literal("radar").nullable(),
-    })
-  ),
-});
-
-export const radarCatalogTool: McpToolDefinition<
-  typeof radarCatalogInput,
-  typeof radarCatalogOutput
-> = {
-  name: "omniroute_radar_catalog",
-  description:
-    "Reads the local signed Radar catalog with optional provider and curated-family filters. Never syncs or writes data.",
-  inputSchema: radarCatalogInput,
-  outputSchema: radarCatalogOutput,
-  scopes: ["read:radar"],
-  auditLevel: "none",
-  phase: 1,
-  sourceEndpoints: ["/api/radar/catalog"],
 };
 
 // --- Tool 10: omniroute_web_search ---
