+++ open-sse/mcp-server/server.ts
@@ -17,8 +17,6 @@ import {
   routeRequestInput,
   costReportInput,
   listModelsCatalogInput,
-  radarCatalogInput,
-  radarCatalogOutput,
   webSearchInput,
   webFetchInput,
   simulateRouteInput,
@@ -95,9 +93,9 @@ import { normalizeQuotaResponse } from "../../src/shared/contracts/quota.ts";
 import { resolveOmniRouteBaseUrl } from "../../src/shared/utils/resolveOmniRouteBaseUrl.ts";
 import { sanitizeErrorMessage } from "../utils/error.ts";
 import { getMcpModelsCatalog } from "./catalog.ts";
-import { getMcpRadarCatalog } from "./radarCatalog.ts";
+import { registerRadarCatalogTool } from "./radarCatalog.ts";
+import type { TextToolResult } from "./toolResult.ts";
 export { getMcpModelsCatalog } from "./catalog.ts";
-export { getMcpRadarCatalog } from "./radarCatalog.ts";
 
 const OMNIROUTE_BASE_URL = resolveOmniRouteBaseUrl();
 const MCP_ENFORCE_SCOPES = process.env.OMNIROUTE_MCP_ENFORCE_SCOPES === "true";
@@ -149,11 +147,6 @@ function readMcpAccessibilityConfig(): McpAccessibilityConfig {
     return { ...DEFAULT_MCP_ACCESSIBILITY_CONFIG };
   }
 }
-
-type TextToolResult = {
-  content: Array<{ type: "text"; text: string }>;
-  isError?: boolean;
-};
 
 function toRecord(value: unknown): JsonRecord {
   return value && typeof value === "object" && !Array.isArray(value) ? (value as JsonRecord) : {};
@@ -574,29 +567,6 @@ async function handleListModelsCatalog(args: { provider?: string; capability?: s
   }
 }
 
-async function handleRadarCatalog(args: {
-  provider?: string;
-  familyId?: string;
-  enabledOnly: boolean;
-}) {
-  const start = Date.now();
-  try {
-    const result = radarCatalogOutput.parse(await getMcpRadarCatalog(args));
-    await logToolCall(
-      "omniroute_radar_catalog",
-      args,
-      { modelCount: result.models.length },
-      Date.now() - start,
-      true
-    );
-    return { content: [{ type: "text" as const, text: JSON.stringify(result, null, 2) }] };
-  } catch (error) {
-    const message = sanitizeErrorMessage(error) || "Failed to read Radar catalog";
-    await logToolCall("omniroute_radar_catalog", args, null, Date.now() - start, false, message);
-    return { content: [{ type: "text" as const, text: `Error: ${message}` }], isError: true };
-  }
-}
-
 async function handleWebSearch(args: {
   query: string;
   max_results?: number;
@@ -842,16 +812,7 @@ export function createMcpServer(): McpServer {
     )
   );
 
-  server.registerTool(
-    "omniroute_radar_catalog",
-    {
-      description: "Reads the local signed Radar catalog with optional provider and family filters",
-      inputSchema: radarCatalogInput,
-    },
-    withScopeEnforcement("omniroute_radar_catalog", (args) =>
-      handleRadarCatalog(radarCatalogInput.parse(args))
-    )
-  );
+  registerRadarCatalogTool(server, withScopeEnforcement);
 
   server.registerTool(
     "omniroute_simulate_route",
