+++ open-sse/mcp-server/toolResult.ts
@@ -1,0 +1,4 @@
+export type TextToolResult = {
+  content: Array<{ type: "text"; text: string }>;
+  isError?: boolean;
+};
